backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai
python-dotenv
google-generativeai
